s3-image-uploader/requirements.txt
psycopg2-binary==2.9.1
python-dotenv==0.19.1
boto3==1.34.98
requests==2.26.0